#180 - Split out pdf box output device to slow and fast versions.
+ Renames PdfBoxOutputDevice to PdfBoxSlowOutputDevice.
+ Creates interface PdfBoxOutputDevice extracted from PdfBoxSlowOutputDevice.
+ Copies PdfBoxSlowOutputDevice to PdfBoxFastOutputDevice.
Now we can work on PdfBoxFastOutputDevice without risk of messing up the old slow renderer.
